Add files via upload
I developed a simple calculator using Html ,Css ,Javascript as a part of the web development internship at codsoft. In this task i have created aq simple calculator which perform the basic arithmetic  operations such as addition, subtraction.the key features is that this calculator shows the calculation history of all the operations you have performed.