Add files via upload
As a web deve;opment intern at codsoft i have completed my second task of creating a simple landing page.In this task i have created a landing page for "The Exclusive Chocolates"  which visually appealing chocolates images and different varities of flavour offered with the price details. I had  created a customers reviews form for the review of the customers to be gathered.